Activate beartype claw on lcm.regime_building
Adds `lcm.regime_building` to the import-claw registrations in
`lcm/__init__.py`, mapping type violations to `ModelInitializationError`
(regime compilation is part of model construction).

Wraps `dags.with_signature` / `dags.rename_arguments` via a thin
helper in `lcm.utils._dags_forwarders` that defaults
`forwarder=True`. Each direct caller of those dags helpers inside
pylcm produces a generic `*args, **kwargs` forwarder whose annotations
describe the inner function's contract, not the wrapper's own call
protocol. `forwarder=True` advertises the wrapper as a permissive
forwarder on its `__annotations__`, so beartype's claw treats it as
universally permissive and skips per-parameter enforcement — matching
the wrapper's actual runtime behaviour. dags' own `get_annotations`
recovers the user-described view via its existing args/kwargs-mismatch
fallback.

Pins `dags` to the `feat/no-type-check-flag` branch (PR
OpenSourceEconomics/dags#82) which adds the `forwarder` flag. Will be
replaced with a released version once that PR lands.

Annotation drift fixed alongside activation:

- `collect_state_transitions(states: ...)` widened to
  `Mapping[StateName, Grid | None]`; test mocks pass `None` for
  placeholder states.
- `map_coordinates(coordinates: ...)` widened to
  `Sequence[Array] | Array`; callers pass a 2D `jnp.array` (a single
  Array, not a sequence) and JAX produces a single tracer under vmap.
- `_get_weights_func_for_shock.weights_func_runtime.shock_kw` typed as
  `dict[str, float | FloatND]`; under JIT the runtime shock params
  arrive as tracers.
- `solve_brute.solve.running_any_nan` / `running_any_inf` typed as
  `BoolND` to match the underlying `jnp.zeros((), dtype=bool)`.
- `diagnostics._wrap_with_reduction.reduced.**kwargs` typed as
  `Array | Mapping[str, Array]` since `next_regime_to_V_arr` flows
  through it as a mapping alongside Array-valued state/action inputs.
- One test fixture (`tests/test_next_state.py::test_create_stochastic_next_func`)
  updated to pass an `int32`-typed `labels` array.

Out of scope: extending the claw to `lcm.solution` / `lcm.simulation`,
which surfaces further annotation drift the claw correctly catches but
that needs its own pass.

+++ tests/test_next_state.py
@@ -102,7 +102,7 @@ def test_get_next_state_function_with_simulate_target():
 
 
 def test_create_stochastic_next_func():
-    labels = jnp.arange(2)
+    labels = jnp.arange(2, dtype=jnp.int32)
     got_func = _create_discrete_stochastic_next_func(
         target="t", next_state_name="next_a", labels=labels
     )
